refactor: move DebugApi in a dedicated `debug` package
BREAKING CHANGE: DebugApi must now be imported from `io.gravitee.definition.model.debug.DebugApi`
