+++ jackson/src/main/java/io/gravitee/definition/jackson/datatype/api/deser/DebugApiDeserializer.java
@@ -21,9 +21,9 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.deser.std.StdScalarDeserializer;
 import io.gravitee.definition.model.Api;
-import io.gravitee.definition.model.DebugApi;
 import io.gravitee.definition.model.HttpRequest;
 import io.gravitee.definition.model.HttpResponse;
+import io.gravitee.definition.model.debug.DebugApi;
 import java.io.IOException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
